+++ tests/playwright/shiny/bugs/1345-render-data-frame-input/app.py
@@ -22,7 +22,7 @@ def server(input: Inputs):
 
     @render.text
     def cell_selection():
-        cell_selection = df1.input_cell_selection()
+        cell_selection = df1.cell_selection()
         if cell_selection is None:
             return "No cells selected"
         if cell_selection["type"] != "row":